library/pyproject.toml
[tool.poetry]
name = "testlib"
version = "0.1.0"
description = ""
authors = ["Chris Mamon <chrisam1993@live.com>"]
readme = "README.md"
packages = [{include = "testlib", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
typing-extensions = "^4.7.1"
pyspark = "^3.4.1"
pandas = "^2.0.3"
boto3-stubs = {extras = ["essential"], version = "^1.28.19"}
delta-spark = "^2.4.0"
dbt-spark = {extras = ["pyhive"], version = "^1.6.0"}
pytest-order = "^1.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.282"
pytest = "^7.4.0"
moto = {extras = ["s3"], version = "^4.1.14"}
flask = "^2.3.2"
flask-cors = "^4.0.0"
dbt-spark = {extras = ["session"], version = "^1.6.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
